Fix some of the observable sharing; defaults
Default the model and intent better. Provide a better starting point for the shareValue.

Still trying to figure out why things are wiring up in such a way that I need delays for the default model and also when passing in an observable of a single gear. Is this a signal that I may need to rewrite this with lower level Rx concepts than the operators?