:baseball: Add a model catch to gears
Promotes the idea of a catch directly on top of the model (between the
model and its teeth) to a first class citizen. This should make it much
easier to keep from having "stuck transmissions".

The default catch is simply a throw, so behavior should be the same, but
I will bump semver major just in case.

Also, cleaned up a bunch of lets to consts and added a tslint config.